Send PHPUnit Coverage Listender to the cemetary
This hasn't been touched in a while, and I think I mentioned this to @toopay a while back. Three years should be considered a reasonable cut-off for a project. If anyone wants to bring it back from the [Cemetary](https://github.com/thephpleague/thephpleague.github.io/wiki/Cemetery) they can.