Improves wording of quality rule "changelogs"
The German wording of "Behaltet ein Änderungsprotokoll" is a 1:1 translation of "Keep a changelog". However, this is not what we'd say in Germany. This way it means "Do not drop it", but it does not mean "keep it up to date". The verb "führen" would fit better but the whole sentence would stil be a bit weird.

My replacement means something like "Track changes in a changelog" and is a way more natural way of saying it in German. I've also replaced the word "Änderungsprotokoll", which is a translation of "Changelog". Everbody in IT knows what a changelog is, there is no need to translate it. However, I've also changed the keepachangelog.com link to the German version of the page.